requirements.txt
lancedb
openai
tree-sitter==0.21.3
tree_sitter_languages
pandas
tiktoken
rerankers
rerankers[transformers]
python-dotenv
fastapi
uvicorn[standard]
pydantic
python-multipart
aiofiles
requests
nest_asyncio
voyageai
numpy
concurrent-log-handler
tqdm
openai-agents
multilspy
